backend/requirements.txt
# Core frameworks
Flask==2.2.3
Flask-Cors==3.0.10
Werkzeug==2.2.3

# Machine learning & image processing
numpy==1.24.2
scikit-learn==1.2.2
scikit-image==0.20.0
opencv-python-headless==4.7.0.72
Pillow==9.4.0

# HTTP & API tools
requests==2.28.2
urllib3==1.26.15

# Production server
gunicorn==20.1.0

# Testing & development
pytest==7.3.1
